Implement query correction...
1. In the web page, when a user is typing a query, the page will send
the unfinished query to the server with a certain intervals. a). If the
last term of the query exists in the dictionary comes from the index,
then several expanded query will be sent back to the pages as
suggestions. b). If the last term does not exist, a spell check will be
ran and several corrected query will be sent back to the page….